fix: merge fake session clone helpers to satisfy pkg-file-count
